finished refactoring store to name 'products'
