+++ client/components/navbar.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import {connect} from 'react-redux'
 import {Link} from 'react-router-dom'
 import {logout} from '../store'
-import {fetchProducts} from '../store/allProducts'
+import {fetchProducts} from '../store/products'
 import './navbar.css'
 
 const Navbar = ({handleClick, isLoggedIn, getProductsFromServer}) => (
