+++ client/routes.js
@@ -8,7 +8,6 @@ import SingleProduct from './components/single-product'
 import AllProducts from './components/all-products'
 import MyAccount from './components/my-account'
 import AllReviews from './components/all-reviews'
-import SearchResults from './components/search-results'
 import Cart from './components/cart'
 import ReturnPolicy from './components/return-policy'
 import Homepage from './components/homepage'
